amd64: enable at_wini IDE driver, start it in ramdisk rc
Previously the amd64 ramdisk only started virtio_blk and only when QEMU
was configured to expose a virtio device.  With the default QEMU
`-cdrom` (PIIX3 IDE attachment), no block driver came up and `cdprobe`
in the boot rc hung forever.

Build at_wini for amd64:
- minix/drivers/storage/Makefile: include at_wini in the build for both
  i386 and x86_64.
- minix/drivers/storage/at_wini/at_wini.c: the DMAERR macro passed a
  pointer difference (ptrdiff_t -> long on amd64) to a `%d` format
  string.  Cast to int so -Werror doesn't reject the amd64 build.

Package and start it:
- minix/drivers/storage/ramdisk/Makefile: include at_wini in PROGRAMS
  for amd64 so the ramdisk gets the binary.
- minix/drivers/storage/ramdisk/proto: install /service/at_wini in the
  amd64 boot ramdisk.
- minix/drivers/storage/ramdisk/rc: when virtio_blk isn't selected
  (typical for plain `-cdrom file.iso`), fall back to starting at_wini
  on /dev/c0d0 (and c1d0 as instance 1) so the IDE CD-ROM is reachable
  before cdprobe runs.
- distrib/sets/lists/minix-base/md.amd64: add /service/at_wini to the
  amd64 base set.

With this in place the boot rc's `cdprobe` finds the CD, VFS mounts
the root filesystem (`fs_c0d2` at /), and init runs further than
before — bringing up 12+ services vs. the 6 that came up before.

+++ minix/drivers/storage/at_wini/at_wini.c
@@ -965,7 +965,7 @@ static int error_dma(const struct wini *wn)
 
 #define DMAERR(msg) \
 	printf("at_wini%ld: bad DMA: %s. Disabling DMA for drive %d.\n",	\
-		w_instance, msg, wn - wini);				\
+		w_instance, msg, (int)(wn - wini));			\
 	printf("at_wini%ld: workaround: set %s=1 in boot monitor.\n", \
 		w_instance, NO_DMA_VAR); \
 	return 1;	\
